requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.111.1
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ── Database (PostgreSQL async) ───────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9        # needed by Alembic for sync migrations

# ── Redis (caching + token denylist) ─────────────────────────────────────────
redis[hiredis]==5.0.4

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0            # argon2 password hashing (used in security.py)

# ── Validation / config ───────────────────────────────────────────────────────
pydantic[email]==2.7.1
pydantic-settings==2.3.1
python-dotenv==1.0.1

# ── File parsing ──────────────────────────────────────────────────────────────
PyMuPDF==1.24.5
python-docx==1.1.2

# ── Background tasks ──────────────────────────────────────────────────────────
celery[redis]==5.4.0

# ── ML (optional) ─────────────────────────────────────────────────────────────
scikit-learn==1.5.0
numpy==1.26.4
joblib==1.4.2
